Implement enhancement #8
